+++ cx-portal-shared-components/src/components/basic/Dropzone/types.ts
@@ -1,0 +1,25 @@
+export type UploadFileStatus =
+  | 'new'
+  | 'uploading'
+  | 'upload_success'
+  | 'upload_error'
+
+export type UploadFile = {
+  name: string
+  size: number
+  status: UploadFileStatus
+  progressPercent?: number
+}
+
+export type DropZoneDropAreaTranslations = {
+  title: string | JSX.Element
+  subTitle: string | JSX.Element
+  errorTitle: string | JSX.Element
+}
+
+export type DropZonePreviewTranslations = {
+  placeholder: string | JSX.Element
+  uploadProgess: string
+  uploadSuccess: string | JSX.Element
+  uploadError: string | JSX.Element
+}