+++ cx-portal/src/types/Config.tsx
@@ -65,6 +65,8 @@ import IDPManagement from 'components/pages/IDPManagement'
 import IDPDetail from 'components/pages/IDPDetail'
 import AppReleaseProcessForm from 'components/pages/AppReleaseProcess/components'
 import CompanyRoles from 'components/pages/CompanyRoles'
+import UseCase from 'components/pages/UseCase'
+import Deactivate from 'components/pages/AppOverview/Deactivate'
 
 /**
  * ALL_PAGES
@@ -329,6 +331,21 @@ export const ALL_PAGES: IPage[] = [
   { name: PAGES.INTRODUCTION_CONFIRMITY_BODY, element: <CompanyRoles /> },
   { name: PAGES.INTRODUCTION_PARTICIPANT, element: <CompanyRoles /> },
   { name: PAGES.INTRODUCTION_SERVICE_PROVIDER, element: <CompanyRoles /> },
+  { name: PAGES.USE_CASE, element: <UseCase /> },
+  { name: PAGES.USE_CASE_TRACABILITY, element: <UseCase /> },
+  {
+    name: PAGES.DEACTIVATE,
+    isRoute: true,
+    element: (
+      <Route
+        key={PAGES.DEACTIVATE}
+        path={PAGES.DEACTIVATE}
+        element={<Deactivate />}
+      >
+        <Route path=":appId" element={<Deactivate />} />
+      </Route>
+    ),
+  },
 ]
 
 export const ALL_OVERLAYS: IOverlay[] = [
@@ -466,6 +483,10 @@ export const mainMenuFullTree = [
       { name: PAGES.INTRODUCTION_SERVICE_PROVIDER, hint: HINTS.NEW },
       { name: PAGES.INTRODUCTION_CONFIRMITY_BODY, hint: HINTS.NEW },
     ],
+  },
+  {
+    name: PAGES.USE_CASE,
+    children: [{ name: PAGES.USE_CASE_TRACABILITY, hint: HINTS.NEW }],
   },
   {
     name: PAGES.MARKETPLACE,